Module slothjs::ast::stmt

AST definitions for JavaScript statements.

The main struct here is Statement, which wraps Stmt enum.

Structs

BlockStatement

BreakStatement

CatchClause

ContinueStatement

ExpressionStatement

ForInStatement

ForStatement

FunctionDeclaration

IfStatement

LabelStatement

ReturnStatement

Statement

Statement represents an Stmt together with its source span, if any.

SwitchCase

SwitchStatement

ThrowStatement

TryStatement

VariableDeclaration

VariableDeclarator

Enums

DeclarationKind

ForInTarget

Stmt

Functions

block

make a BlockStatement

expr

make an ExpressionStatement

return_

make a ReturnStatement

var

make a VariableDeclaration from [(var1, value1), …]