Enum slothjs::object::JSValue

pub enum JSValue {
    Undefined,
    Bool(bool),
    Number(JSNumber),
    String(String),
    Ref(JSRef),
}

A JSValue is either a primitive value or a reference to an object.

Variants

Undefined

Bool(bool)

Number(JSNumber)

String(String)

Ref(JSRef)

Implementations

impl JSValue

pub const NULL: JSValue

pub fn to_ref(&self) -> Result<JSRef, Exception>

to_ref() tries to return the underlying object reference, if any. It’s useful for checking if a value points to an object.

pub fn to_json(&self, heap: &Heap) -> Result<JSON, Exception>

pub fn to_string(&self, heap: &mut Heap) -> Result<String, Exception>

to_string() makes a human-readable string representation of the value:

assert_eq!(
    JSValue::from("1").to_string(&mut heap).unwrap(),
    "\"1\""
);
assert_eq!(
    JSValue::from(1).to_string(&mut heap).unwrap(),
    "1"
);

let json_object = json!({"one": 1});
let example_object = heap.object_from_json(&json_object);
assert_eq!(
    example_object.to_string(&mut heap).unwrap(),
    "{ one: 1 }"
);

let json_array = json!([1, 2]);
let example_array = heap.object_from_json(&json_array);
assert_eq!( example_array.to_string(&mut heap).unwrap(), "[1, 2]" );

pub fn stringify(&self, heap: &mut Heap) -> Result<String, Exception>

stringify() makes everything into a string used for evaluation in a string context. It corresponds to .toString() in JavaScript

pub fn numberify(&self, heap: &Heap) -> Option<JSNumber>

numberify() tries to make everything into a numeric value for evalation in a numeric context. It is slightly more strict than +value in JavaScript: only value.numberify().unwrap_or(f64::NAN) corresponds to +value in JavaScript.

pub fn boolify(&self, heap: &Heap) -> bool

boolify() treats everythings as a truthy value. ES5: ToBoolean

pub fn objectify(&self, heap: &mut Heap) -> JSRef

objectify() wraps a primitive into its object:

undefined becomes null

bool/number/string becomes Boolean/Number/String

objects just return their reference.

pub fn type_of(&self, heap: &Heap) -> &'static str

Javascript’s typeof

pub fn loose_eq(&self, other: &JSValue, heap: &Heap) -> bool

Abstract Equality Comparison, ==: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Equality_comparisons_and_sameness#Loose_equality_using_

pub fn strict_eq(&self, other: &JSValue, _heap: &Heap) -> bool

Strict Equality, === https://developer.mozilla.org/en-US/docs/Web/JavaScript/Equality_comparisons_and_sameness#strict_equality_using

pub fn numerically<F>(&self, other: &JSValue, heap: &Heap, op: F) -> JSValue where
F: Fn(f64, f64) -> f64,

pub fn plus(
&self,
other: &JSValue,
heap: &mut Heap
) -> Result<JSValue, Exception>

Addition operator: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Addition

pub fn minus(&self, other: &JSValue, heap: &Heap) -> Result<JSValue, Exception>

Subtraction operator: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Subtraction

pub fn compare<StrCmpFn: Fn(&str, &str) -> bool, NumCmpFn: Fn(f64, f64) -> bool>(
&self,
other: &JSValue,
heap: &Heap,
stringly: StrCmpFn,
numberly: NumCmpFn
) -> JSValue