Struct slothjs::object::JSObject

pub struct JSObject {
    pub proto: JSRef,
    pub value: ObjectValue,
    pub properties: HashMap<String, Property>,
}

Javascript objects. A JSObject always has a proto. It can have an optional ObjectValue (a primitive or array/function/closure). It has a dictionary of properties.

Fields
proto: JSRefvalue: ObjectValueproperties: HashMap<String, Property>
Implementations

impl JSObject

pub fn new() -> JSObject

pub fn from_func(f: NativeFunction) -> JSObject

Wrap the given native call into a Function.

pub fn from_closure(closure: Closure) -> JSObject

Wrap the given closure into a Function.

pub fn from_array(values: Vec<JSValue>) -> JSObject

Wrap the given vector into an Array.

pub fn from_bool(value: bool) -> JSObject

pub fn to_primitive(&self, _heap: &Heap) -> Option<JSValue>

It’s roughly Object.valueOf(self)

pub fn as_array(&self) -> Option<&JSArray>

It self is an Array, give its underlying storage.

pub fn as_array_mut(&mut self) -> Option<&mut JSArray>

It self is an Array, give its underlying storage mutably.

pub fn get_value(&self, name: &str) -> Option<&JSValue>

Tries to get JSValue of the own property name. This might call getters of the property.

pub fn lookup_value<'a>(
&'a self,
name: &str,
heap: &'a Heap
) -> Option<&'a JSValue>

Check own and all inherited properties for name and returns the first found value.

pub fn protochain<'a>(&self, heap: &'a Heap) -> ProtoChainIter<'a>

pub fn set(
&mut self,
name: &str,
content: Content,
access: Access
) -> Result<(), Exception>

If the own property name does not exist, create it with the given content and access. Otherwise, if name is a number and self is an Array, assign the value of content into the array. Otherwise:

if the existing own property is not configurable and the given access differs, fail.

if the existing own property is not writable, fail

replace content and access of the property.

pub fn set_even_nonwritable(
&mut self,
name: &str,
content: Content,
access: Access
) -> Result<(), Exception>

Just like .set(), but updates even readonly properties.

pub fn update(&mut self, name: &str, value: JSValue) -> Result<(), Exception>

If name is a number and self is an Array, just set the array elemnt to value. Otherwise: if the own property name does not exist, create it with Access::all() and set to Content::from(value). If the own property exists already, call .set() with its current access. This will fail to update non-writable properties.

pub fn update_even_nonwritable(
&mut self,
name: &str,
value: JSValue
) -> Result<(), Exception>

Just like .update(), but updates even non-writable properties.

pub fn set_property(
&mut self,
name: &str,
content: Content
) -> Result<(), Exception>

pub fn set_system(
&mut self,
name: &str,
content: Content
) -> Result<(), Exception>

pub fn set_hidden(
&mut self,
name: &str,
content: Content
) -> Result<(), Exception>

pub fn set_nonconf(
&mut self,
name: &str,
content: Content
) -> Result<(), Exception>

pub fn set_readonly(
&mut self,
name: &str,
content: Content
) -> Result<(), Exception>

pub fn to_json(&self, heap: &Heap) -> Result<JSON, Exception>

Create a JSON from this JSObject.

pub fn to_string(&self, heap: &mut Heap) -> Result<String, Exception>

Create a human-readable representation of contents of an Array or an Object.