Trait slothjs::SourceNode

pub trait SourceNode: Sized {
    fn to_error(&self) -> JSON;
    fn get_location(&self) -> Option<Location>;
    fn get_literal(&self, property: &str) -> Result<Literal, ParseError>;
    fn get_bool(&self, property: &str) -> Result<bool, ParseError>;
    fn get_str(&self, property: &str) -> Result<JSString, ParseError>;
    fn map_node<T, F>(&self, property: &str, action: F) -> Result<T, ParseError>
    where
        F: FnMut(&Self) -> Result<T, ParseError>;
    fn map_array<T, F>(
        &self,
        property: &str,
        func: F
    ) -> Result<Vec<T>, ParseError>
    where
        F: FnMut(&Self) -> Result<T, ParseError>;

    fn expect_str(
        &self,
        property: &str,
        value: &'static str
    ) -> Result<(), ParseError> { ... }
    fn map_opt_node<T, F>(
        &self,
        property: &str,
        action: F
    ) -> Result<Option<T>, ParseError>
    where
        F: FnMut(&Self) -> Result<T, ParseError>,
    { ... }
}

SourceNode is how ParseFrom::parse_from sees AST nodes.

Required Methods

fn to_error(&self) -> JSON

Location of the node where an error happened.

fn get_location(&self) -> Option<Location>

Try to get source mapping for self.

fn get_literal(&self, property: &str) -> Result<Literal, ParseError>

Use the node as a literal.

fn get_bool(&self, property: &str) -> Result<bool, ParseError>

Get the boolean value of a child node with name property. It’s a ParseError if it does not exist or does not have a boolean meaning.

fn get_str(&self, property: &str) -> Result<JSString, ParseError>

Get the string value of a child node with name property. It’s a ParseError if it does not exist or does not have a string meaning.

fn map_node<T, F>(&self, property: &str, action: F) -> Result<T, ParseError> where
F: FnMut(&Self) -> Result<T, ParseError>,

Get a child node with this name; if it does not exist, return None. Then transform it through action, propagating its Result out. A child node exis

fn map_array<T, F>(&self, property: &str, func: F) -> Result<Vec<T>, ParseError> where
F: FnMut(&Self) -> Result<T, ParseError>,

Map the array of children of a child node with name property. It’s a ParseError if it does not exist or does not have an array meaning.

Provided Methods

fn expect_str(
&self,
property: &str,
value: &'static str
) -> Result<(), ParseError>

Check that the value of property is a string equal to value. Depends on SourceNode::get_str.

fn map_opt_node<T, F>(
&self,
property: &str,
action: F
) -> Result<Option<T>, ParseError> where
F: FnMut(&Self) -> Result<T, ParseError>,