Struct slothjs::heap::Heap

pub struct Heap {
    pub loc: Option<Box<Location>>,
    // some fields omitted
}

Runtime heap

Fields
loc: Option<Box<Location>>
Implementations

impl Heap

pub const NULL: JSRef

pub const GLOBAL: JSRef

pub const OBJECT_PROTO: JSRef

pub const FUNCTION_PROTO: JSRef

pub const ARRAY_PROTO: JSRef

pub const BOOLEAN_PROTO: JSRef

pub const ERROR_PROTO: JSRef

pub fn new() -> Self

pub fn get(&self, objref: JSRef) -> &JSObject

pub fn get_mut(&mut self, objref: JSRef) -> &mut JSObject

pub fn alloc(&mut self, object: JSObject) -> JSRef

pub fn object_from_json(&mut self, json: &JSON) -> JSValue

Deserializes JSON into objects on the heap

pub fn interpret_this(&mut self) -> Result<Interpreted, Exception>

Find out what this currently is.

pub fn scope_mut(&mut self) -> &mut JSObject

pub fn declare<'a>(
&mut self,
variables: impl Iterator<Item = &'a Identifier>,
functions: impl Iterator<Item = &'a FunctionDeclaration>
) -> Result<(), Exception>

pub fn lookup_var(&mut self, name: &str) -> Option<Interpreted>

pub fn lookup_path(&self, names: &[&str]) -> Result<Interpreted, Exception>

Lookup a property chain starting from the current scope, e.g.

assert_eq!(
    heap.lookup_path(&["Array", "prototype"]).unwrap(),
    Interpreted::from(Heap::ARRAY_PROTO)
);

pub fn enter_new_scope<T, F>(
&mut self,
this_ref: JSRef,
captured_scope: JSRef,
action: F
) -> Result<T, Exception> where
F: FnMut(&mut Heap) -> Result<T, Exception>,

pub fn lookup_protochain(
&self,
objref: JSRef,
propname: &str
) -> Option<Interpreted>

Find the location of propname on the prototype chain of objref. Return None or Some(Interpreted::Member{..}) pointing to the found own property.

pub fn execute(
&mut self,
func_ref: JSRef,
call: CallContext
) -> Result<Interpreted, Exception>

Given a func_ref to a closure or a native call and a set of arguments, executes the function. this_ref is bound as this.

pub fn throw<T>(&self, exc: Exception) -> Result<T, Exception>