Trait slothjs::SourceNode

pub trait SourceNode: Sized + Clone {
    fn get_literal(&self) -> Literal;

    fn get_location(&self) -> Option<Location>;

    fn get_node(
        &self, 
        property: &'static str
    ) -> Result<&Self, ParseError<Self>>;

    fn map_node<T, F>(
        &self, 
        property: &'static str, 
        action: F
    ) -> Result<Option<T>, ParseError<Self>>
    where
        F: FnMut(&Self) -> Result<T, ParseError<Self>>;

    fn get_bool(&self, property: &'static str) -> Result<bool, ParseError<Self>>;

    fn get_str(&self, property: &'static str) -> Result<&str, ParseError<Self>>;

    fn get_array(
        &self, 
        property: &'static str
    ) -> Result<&[Self], ParseError<Self>>;

    fn expect_str(
        &self, 
        property: &'static str, 
        value: &'static str
    ) -> Result<(), ParseError<Self>>;
}

SourceNode is how ParseFrom::parse_from sees AST nodes.

Required methods

fn get_literal(&self) -> Literal

Make the node into a literal.

fn get_location(&self) -> Option<Location>

Try to get source mapping for self.

fn get_node(&self, property: &'static str) -> Result<&Self, ParseError<Self>>

Get a child node with this name; it’s a ParseError if it does not exist.

fn map_node<T, F>(
&self,
property: &'static str,
action: F
) -> Result<Option<T>, ParseError<Self>> where
F: FnMut(&Self) -> Result<T, ParseError<Self>>,

Get a child node with this name; if it does not exist, return None. Then transform it through action, propagating its Result out.

fn get_bool(&self, property: &'static str) -> Result<bool, ParseError<Self>>

Get the boolean value of a child node with name property. It’s a ParseError if it does not exist or does not have a boolean meaning.

fn get_str(&self, property: &'static str) -> Result<&str, ParseError<Self>>

Get the string value of a child node with name property. It’s a ParseError if it does not exist or does not have a string meaning.

fn get_array(&self, property: &'static str) -> Result<&[Self], ParseError<Self>>

Get the array of children of a child node with name property. It’s a ParseError if it does not exist or does not have an array meaning.

fn expect_str(
&self,
property: &'static str,
value: &'static str
) -> Result<(), ParseError<Self>>

Check that the value of property is a string equal to value.

Loading content...